+++ lib/SignInPages/company_signup.dart
@@ -80,22 +80,17 @@ class _CompanySignUpState extends State<CompanySignUp> {
       };
 
       // Shared Preference implementations
-      SharedPref.saveCompanyNameSharedPreference(
-          companyNameTextEditingController.text);
+      SharedPref.saveCompanyNameSharedPreference(companyNameTextEditingController.text);
       SharedPref.saveEmailSharedPreference(emailTextEditingController.text);
       SharedPref.saveIsCompanySharedPreference(true);
       SharedPref.saveLoggedInSharedPreference(true);
       if (logoUrl != '') SharedPref.saveImgUrlSharedPreference(logoUrl);
-      SharedPref.saveCompanyDescriptionSharedPreference(
-          descriptionTextEditingController.text);
-      SharedPref.saveCompanyServiceTypeSharedPreference(
-          CompanyConstants.serviceType);
+      SharedPref.saveCompanyDescriptionSharedPreference(descriptionTextEditingController.text);
+      SharedPref.saveCompanyServiceTypeSharedPreference(CompanyConstants.serviceType);
       SharedPref.saveLogoUrlSharedPreference(logoUrl);
 
-      authMethods
-          .signUpWithEmailAndPassword(emailTextEditingController.text,
-              passwordTextEditingController.text)
-          .then((value) {
+      authMethods.signUpWithEmailAndPassword
+        (emailTextEditingController.text, passwordTextEditingController.text).then((value) {
         CompanyConstants.companyName = companyNameTextEditingController.text;
         CompanyConstants.description = descriptionTextEditingController.text;
         CompanyConstants.email = emailTextEditingController.text;
@@ -120,176 +115,177 @@ class _CompanySignUpState extends State<CompanySignUp> {
   // UI of the Page
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-        resizeToAvoidBottomPadding: false,
-        body: Container(
-          width: double.infinity,
-          height: double.infinity,
-          padding: EdgeInsets.symmetric(vertical: 64, horizontal: 16),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Container(
-                height: 100,
-                width: MediaQuery.of(context).size.width - 30,
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.end,
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      "Create Services Account,",
-                      style:
-                          TextStyle(fontSize: 27, fontWeight: FontWeight.bold),
-                    ),
-                    Text(
-                      "Sign up to get started!",
-                      style: TextStyle(
-                          fontSize: 20,
-                          fontWeight: FontWeight.bold,
-                          color: Colors.grey),
-                    ),
-                  ],
-                ),
-              ),
-              Container(
-                height: MediaQuery.of(context).size.height / 1.8,
-                width: MediaQuery.of(context).size.width - 30,
-                child: Form(
-                  key: formKey,
+    return Scaffold(        
+        body: SingleChildScrollView(
+          child: Container(
+            width: MediaQuery.of(context).size.width,
+            height: MediaQuery.of(context).size.height,
+            padding: EdgeInsets.symmetric(vertical: 64, horizontal: 16),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.stretch,
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                Container(
+                  height: 100,
+                  width: MediaQuery.of(context).size.width - 30,
                   child: Column(
-                    mainAxisAlignment: MainAxisAlignment.spaceAround,
+                    mainAxisAlignment: MainAxisAlignment.end,
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
-                      Padding(
-                        padding: const EdgeInsets.only(bottom: 16),
-                        child: Row(
-                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                          children: [
-                            _image == null
-                                ? Image.asset(
-                                    'assets/icons/noImg.png',
-                                    height: 100,
-                                    width: 100,
-                                  )
-                                : Container(
-                                    decoration: BoxDecoration(
-                                        color: Colors.grey[300],
-                                        borderRadius:
-                                            BorderRadius.circular(50)),
-                                    child: ClipRRect(
-                                        borderRadius: BorderRadius.circular(50),
-                                        child: Image.file(
-                                          _image,
-                                          width: 100,
-                                          height: 100,
-                                        ))),
-                            RaisedButton(
-                              onPressed: () async {
-                                await getImage();
-                              },
-                              color: const Color(0xFFCA436B),
-                              splashColor: Colors.white.withOpacity(0.6),
-                              elevation: 10,
-                              child: Text(
-                                'Add a Company Logo',
-                                style: TextStyle(color: Colors.white),
-                              ),
-                            )
-                          ],
-                        ),
+                      Text(
+                        "Create Services Account,",
+                        style:
+                            TextStyle(fontSize: 27, fontWeight: FontWeight.bold),
                       ),
-                      companyNameInput(
-                          context, companyNameTextEditingController),
-                      Container(
-                        padding: EdgeInsets.symmetric(horizontal: 15),
-                        child: Text(
-                          'Select a service type',
-                          textAlign: TextAlign.left,
-                          style: TextStyle(color: Colors.black45),
-                        ),
+                      Text(
+                        "Sign up to get started!",
+                        style: TextStyle(
+                            fontSize: 20,
+                            fontWeight: FontWeight.bold,
+                            color: Colors.grey),
                       ),
-                      serviceTypeInput(context),
-                      descriptionInput(
-                          context, descriptionTextEditingController),
-                      emailInput(context, emailTextEditingController),
-                      passwordInput(
-                          context, passwordTextEditingController, showPassword),
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.start,
-                        children: [
-                          Checkbox(
-                            value: showPassword,
-                            onChanged: (flag) {
-                              setState(() {
-                                showPassword = !showPassword;
-                              });
-                            },
-                            checkColor: Colors.white,
-                            activeColor: Color.fromRGBO(250, 89, 143, 1),
-                          ),
-                          Text("Show Password"),
-                        ],
-                      )
                     ],
                   ),
                 ),
-              ),
-              GestureDetector(
-                onTap: () {
-                  signUp();
-                },
-                child: Container(
-                  height: 60,
+                Container(
+                  height: MediaQuery.of(context).size.height / 1.8,
                   width: MediaQuery.of(context).size.width - 30,
-                  decoration: BoxDecoration(
-                      borderRadius: BorderRadius.circular(50),
-                      boxShadow: [
-                      BoxShadow(
-                          color: darkMode ? Colors.grey[900] : Colors.grey[600],
-                          offset: Offset(4.0, 4.0),
-                          blurRadius: 15.0,
-                          spreadRadius: 1.0),
-                      BoxShadow(
-                        color: darkMode ? Colors.grey[800] : Colors.white,
-                        offset: Offset(-4.0, -4.0),
-                        blurRadius: 15.0,
-                        spreadRadius: 1.0)
-                  ],
-                  gradient: new LinearGradient(
-                    colors: [const Color(0xFF915FB5), const Color(0xFFCA436B)],
-                  )),
-                  child: Center(
-                    child: Text(
-                      "Sign Up",
-                      style: TextStyle(color: Colors.white, fontSize: 20),
+                  child: Form(
+                    key: formKey,
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.spaceAround,
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 16),
+                          child: Row(
+                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                            children: [
+                              _image == null
+                                  ? Image.asset(
+                                      'assets/icons/noImg.png',
+                                      height: 100,
+                                      width: 100,
+                                    )
+                                  : Container(
+                                      decoration: BoxDecoration(
+                                          color: Colors.grey[300],
+                                          borderRadius:
+                                              BorderRadius.circular(50)),
+                                      child: ClipRRect(
+                                          borderRadius: BorderRadius.circular(50),
+                                          child: Image.file(
+                                            _image,
+                                            width: 100,
+                                            height: 100,
+                                          ))),
+                              RaisedButton(
+                                onPressed: () async {
+                                  await getImage();
+                                },
+                                color: const Color(0xFFCA436B),
+                                splashColor: Colors.white.withOpacity(0.6),
+                                elevation: 10,
+                                child: Text(
+                                  'Add a Company Logo',
+                                  style: TextStyle(color: Colors.white),
+                                ),
+                              )
+                            ],
+                          ),
+                        ),
+                        companyNameInput(
+                            context, companyNameTextEditingController),
+                        Container(
+                          padding: EdgeInsets.symmetric(horizontal: 15),
+                          child: Text(
+                            'Select a service type',
+                            textAlign: TextAlign.left,
+                            style: TextStyle(color: Colors.black45),
+                          ),
+                        ),
+                        serviceTypeInput(context),
+                        descriptionInput(
+                            context, descriptionTextEditingController),
+                        emailInput(context, emailTextEditingController),
+                        passwordInput(
+                            context, passwordTextEditingController, showPassword),
+                        Row(
+                          mainAxisAlignment: MainAxisAlignment.start,
+                          children: [
+                            Checkbox(
+                              value: showPassword,
+                              onChanged: (flag) {
+                                setState(() {
+                                  showPassword = !showPassword;
+                                });
+                              },
+                              checkColor: Colors.white,
+                              activeColor: Color.fromRGBO(250, 89, 143, 1),
+                            ),
+                            Text("Show Password"),
+                          ],
+                        )
+                      ],
                     ),
                   ),
                 ),
-              ),
-              Container(
-                height: 70,
-                width: MediaQuery.of(context).size.width - 30,
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Text("I am already a member, "),
-                    GestureDetector(
-                        onTap: () => Navigator.pushReplacement(
-                              context,
-                              PageTransition(
-                                  child: Login(),
-                                  type: PageTransitionType.leftToRightWithFade,
-                                  duration: Duration(milliseconds: 200)),
-                            ),
-                        child: Text(
-                          "Sign In",
-                          style: TextStyle(color: Colors.pink),
-                        )),
-                  ],
+                GestureDetector(
+                  onTap: () {
+                    signUp();
+                  },
+                  child: Container(
+                    height: 60,
+                    width: MediaQuery.of(context).size.width - 30,
+                    decoration: BoxDecoration(
+                        borderRadius: BorderRadius.circular(50),
+                        boxShadow: [
+                        BoxShadow(
+                            color: darkMode ? Colors.grey[900] : Colors.grey[600],
+                            offset: Offset(4.0, 4.0),
+                            blurRadius: 15.0,
+                            spreadRadius: 1.0),
+                        BoxShadow(
+                          color: darkMode ? Colors.grey[800] : Colors.white,
+                          offset: Offset(-4.0, -4.0),
+                          blurRadius: 15.0,
+                          spreadRadius: 1.0)
+                    ],
+                    gradient: new LinearGradient(
+                      colors: [const Color(0xFF915FB5), const Color(0xFFCA436B)],
+                    )),
+                    child: Center(
+                      child: Text(
+                        "Sign Up",
+                        style: TextStyle(color: Colors.white, fontSize: 20),
+                      ),
+                    ),
+                  ),
                 ),
-              ),
-            ],
+                Container(
+                  height: 70,
+                  width: MediaQuery.of(context).size.width - 30,
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      Text("I am already a member, "),
+                      GestureDetector(
+                          onTap: () => Navigator.pushReplacement(
+                                context,
+                                PageTransition(
+                                    child: Login(),
+                                    type: PageTransitionType.leftToRightWithFade,
+                                    duration: Duration(milliseconds: 200)),
+                              ),
+                          child: Text(
+                            "Sign In",
+                            style: TextStyle(color: Colors.pink),
+                          )),
+                    ],
+                  ),
+                ),
+              ],
+            ),
           ),
         ));
   }
