+++ lib/SignInPages/customer_signup.dart
@@ -100,160 +100,161 @@ class _UserSignUpState extends State<UserSignUp> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-        resizeToAvoidBottomPadding: false,
-        body: Container(
-          width: double.infinity,
-          height: double.infinity,
-          padding: EdgeInsets.symmetric(vertical: 64, horizontal: 16),
-          child: Column(
-            // crossAxisAlignment: CrossAxisAlignment.stretch,
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Container(
-                height: 100,
-                width: MediaQuery.of(context).size.width - 30,
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.end,
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      "Create Account,",
-                      style:
-                          TextStyle(fontSize: 30, fontWeight: FontWeight.bold),
-                    ),
-                    Text(
-                      "Sign up to get started!",
-                      style: TextStyle(
-                          fontSize: 20,
-                          fontWeight: FontWeight.bold,
-                          color: Colors.grey),
-                    ),
-                  ],
-                ),
-              ),
-              Container(
-                height: 350,
-                width: MediaQuery.of(context).size.width - 30,
-                // color: Colors.redAccent,
-                child: Form(
-                  key: formKey,
+    return Scaffold(        
+        body: SingleChildScrollView(
+          child: Container(
+            width: MediaQuery.of(context).size.width,
+            height: MediaQuery.of(context).size.height,          
+            padding: EdgeInsets.symmetric(vertical: 64, horizontal: 16),
+            child: Column(
+              // crossAxisAlignment: CrossAxisAlignment.stretch,
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                Container(
+                  height: 100,
+                  width: MediaQuery.of(context).size.width - 30,
                   child: Column(
-                    mainAxisAlignment: MainAxisAlignment.spaceAround,
-                    crossAxisAlignment: CrossAxisAlignment.center,
+                    mainAxisAlignment: MainAxisAlignment.end,
+                    crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
-                      Padding(
-                        padding: const EdgeInsets.only(bottom: 16),
-                        child: Row(
-                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                          children: [
-                            _image == null
-                                ? Image.asset(
-                                    'assets/icons/noImg.png',
-                                    height: 100,
-                                    width: 100,
-                                  )
-                                : Container(
-                                    decoration: BoxDecoration(
-                                        color: Colors.grey[300],
-                                        borderRadius:
-                                            BorderRadius.circular(50)),
-                                    child: ClipRRect(
-                                        borderRadius: BorderRadius.circular(50),
-                                        child: Image.file(
-                                          _image,
-                                          width: 100,
-                                          height: 100,
-                                        ))),
-                            RaisedButton(
-                              onPressed: () async {
-                                await getImage();
-                              },
-                              color: Color.fromRGBO(250, 89, 143, 1),
-                              splashColor: Colors.white.withOpacity(0.6),
-                              elevation: 10,
-                              child: Text(
-                                'Add a Profile Pic',
-                                style: TextStyle(color: Colors.white),
-                              ),
-                            )
-                          ],
-                        ),
+                      Text(
+                        "Create Account,",
+                        style:
+                            TextStyle(fontSize: 30, fontWeight: FontWeight.bold),
                       ),
-                      fullNameInput(context, nameTextEditingController),
-                      emailInput(context, emailTextEditingController),
-                      passwordInput(
-                          context, passwordTextEditingController, showPassword),
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.start,
-                        children: [
-                          Checkbox(
-                            value: showPassword,
-                            onChanged: (flag) {
-                              setState(() {
-                                showPassword = !showPassword;
-                              });
-                            },
-                            checkColor: Colors.white,
-                            activeColor: Color.fromRGBO(250, 89, 143, 1),
-                          ),
-                          Text("Show Password"),
-                        ],
-                      )
+                      Text(
+                        "Sign up to get started!",
+                        style: TextStyle(
+                            fontSize: 20,
+                            fontWeight: FontWeight.bold,
+                            color: Colors.grey),
+                      ),
                     ],
                   ),
                 ),
-              ),
-              RaisedButton(
-                onPressed: () {
-                  print('Trying to sign up');
-                  signUp();
-                },
-                textColor: Colors.white,
-                padding: const EdgeInsets.all(0.0),
-                elevation: 10,
-                shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(20)),
-                child: Container(
-                  height: 60,
-                  width: MediaQuery.of(context).size.width - 40,
-                  decoration: new BoxDecoration(
-                      borderRadius: BorderRadius.circular(20),
-                      gradient: new LinearGradient(colors: [
-                        Color.fromRGBO(250, 89, 143, 1),
-                        Color.fromRGBO(253, 170, 142, 1)
-                      ])),
-                  padding: const EdgeInsets.all(10.0),
-                  child: Center(
-                      child: Text(
-                    "Sign Up",
-                    style: TextStyle(fontSize: 22),
-                  )),
-                ),
-              ),
-              Container(
-                height: 70,
-                width: MediaQuery.of(context).size.width - 30,
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Text("I am already a member, "),
-                    GestureDetector(
-                        onTap: () => Navigator.pushReplacement(
-                              context,
-                              PageTransition(
-                                  child: Login(),
-                                  type: PageTransitionType.leftToRightWithFade,
-                                  duration: Duration(milliseconds: 200)),
+                Container(
+                  height: 350,
+                  width: MediaQuery.of(context).size.width - 30,
+                  // color: Colors.redAccent,
+                  child: Form(
+                    key: formKey,
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.spaceAround,
+                      crossAxisAlignment: CrossAxisAlignment.center,
+                      children: [
+                        Padding(
+                          padding: const EdgeInsets.only(bottom: 16),
+                          child: Row(
+                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                            children: [
+                              _image == null
+                                  ? Image.asset(
+                                      'assets/icons/noImg.png',
+                                      height: 100,
+                                      width: 100,
+                                    )
+                                  : Container(
+                                      decoration: BoxDecoration(
+                                          color: Colors.grey[300],
+                                          borderRadius:
+                                              BorderRadius.circular(50)),
+                                      child: ClipRRect(
+                                          borderRadius: BorderRadius.circular(50),
+                                          child: Image.file(
+                                            _image,
+                                            width: 100,
+                                            height: 100,
+                                          ))),
+                              RaisedButton(
+                                onPressed: () async {
+                                  await getImage();
+                                },
+                                color: Color.fromRGBO(250, 89, 143, 1),
+                                splashColor: Colors.white.withOpacity(0.6),
+                                elevation: 10,
+                                child: Text(
+                                  'Add a Profile Pic',
+                                  style: TextStyle(color: Colors.white),
+                                ),
+                              )
+                            ],
+                          ),
+                        ),
+                        fullNameInput(context, nameTextEditingController),
+                        emailInput(context, emailTextEditingController),
+                        passwordInput(
+                            context, passwordTextEditingController, showPassword),
+                        Row(
+                          mainAxisAlignment: MainAxisAlignment.start,
+                          children: [
+                            Checkbox(
+                              value: showPassword,
+                              onChanged: (flag) {
+                                setState(() {
+                                  showPassword = !showPassword;
+                                });
+                              },
+                              checkColor: Colors.white,
+                              activeColor: Color.fromRGBO(250, 89, 143, 1),
                             ),
-                        child: Text(
-                          "Sign In",
-                          style: TextStyle(color: Colors.pink),
-                        )),
-                  ],
+                            Text("Show Password"),
+                          ],
+                        )
+                      ],
+                    ),
+                  ),
                 ),
-              ),
-            ],
+                RaisedButton(
+                  onPressed: () {
+                    print('Trying to sign up');
+                    signUp();
+                  },
+                  textColor: Colors.white,
+                  padding: const EdgeInsets.all(0.0),
+                  elevation: 10,
+                  shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(20)),
+                  child: Container(
+                    height: 60,
+                    width: MediaQuery.of(context).size.width - 40,
+                    decoration: new BoxDecoration(
+                        borderRadius: BorderRadius.circular(20),
+                        gradient: new LinearGradient(colors: [
+                          Color.fromRGBO(250, 89, 143, 1),
+                          Color.fromRGBO(253, 170, 142, 1)
+                        ])),
+                    padding: const EdgeInsets.all(10.0),
+                    child: Center(
+                        child: Text(
+                      "Sign Up",
+                      style: TextStyle(fontSize: 22),
+                    )),
+                  ),
+                ),
+                Container(
+                  height: 70,
+                  width: MediaQuery.of(context).size.width - 30,
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      Text("I am already a member, "),
+                      GestureDetector(
+                          onTap: () => Navigator.pushReplacement(
+                                context,
+                                PageTransition(
+                                    child: Login(),
+                                    type: PageTransitionType.leftToRightWithFade,
+                                    duration: Duration(milliseconds: 200)),
+                              ),
+                          child: Text(
+                            "Sign In",
+                            style: TextStyle(color: Colors.pink),
+                          )),
+                    ],
+                  ),
+                ),
+              ],
+            ),
           ),
         ));
   }
