+++ lib/views/customer_navigator.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:hack_it_out_demo/helper/sharedpreferences.dart';
 import 'package:hack_it_out_demo/modules/customer_constants.dart';
 import 'package:hack_it_out_demo/views/CustomerPages/customer_account_page.dart';
-import 'package:hack_it_out_demo/views/CustomerPages/customer_chat_page.dart';
 import 'package:hack_it_out_demo/views/CustomerPages/customer_mainpage.dart';
 import 'package:hack_it_out_demo/views/CustomerPages/customer_search_page.dart';
 import 'package:hack_it_out_demo/views/chat/chatroom_list.dart';
